+++ src/pipes/ngSelectValue.pipe.ts
@@ -2,6 +2,7 @@ import {Pipe, PipeTransform} from '@angular/core';
 import {isBlank} from '@jscrpt/common';
 
 import {NgSelectOption} from '../components/option';
+import {DisplayTextFunc} from '../plugins/normalState';
 
 /**
  * Pipe to transform ng select selected option into
@@ -15,8 +16,9 @@ export class NgSelectValuePipe<TValue = any> implements PipeTransform
      * Transforms selected option into string
      * @param options - Selected options to be transformed into text
      * @param nothingSelectedText - Text displayed if nothing is selected
+     * @param optionDisplayText - Function used for transformation of option into display text, defaults to text property of option
      */   
-    public transform(options: NgSelectOption<TValue>|Array<NgSelectOption<TValue>>, nothingSelectedText: string): string
+    public transform(options: NgSelectOption<TValue>|Array<NgSelectOption<TValue>>, nothingSelectedText: string, optionDisplayText: DisplayTextFunc<TValue> = option => option.text): string
     {
         if(isBlank(options) || (Array.isArray(options) && !options.length))
         {
@@ -25,10 +27,10 @@ export class NgSelectValuePipe<TValue = any> implements PipeTransform
 
         if(Array.isArray(options))
         {
-            return options.map(option => option.text).join(', ');
+            return options.map(optionDisplayText).join(', ');
         }
 
-        return options.text;
+        return optionDisplayText(options);
     }
 }